Update README.md
In this assignment, you will be implementing a feed-forward neural network (NN) that will perform classification on the IRIS dataset. The NN will be trained via the stochastic gradient method (SGD), forward propagation and back propagation algorithms. The performance of your implementation will be compared with that of the built-in NN module available in scikit-learn. You will also be asked to answer several questions related to your implementations. For this assignment, we recommend you to attempt the scikit-learn part first. To avoid any potential installation issue, you are encouraged to develop your solution using Google Colab notebooks.

Requirements

In your implementations, please use the function prototype provided (i.e. name of the function, inputs and outputs) in the detailed instructions presented in the remainder of this document. We will be testing your code using a test function that which evokes the provided function prototype. If our testing file is unable to recognize the function prototype you have implemented, you can lose significant portion of your marks. In the assignment folder, the following files are included in thestarter code folder:

NeuralNetwork.py

These files contain the test function and an outline of the functions that you will be implementing. You also need to submit a separate PA2 qa.pdf file that answer questions related to your implementations.

NumPy Implementation Overview

In this specific implementation, the task of the NN is to identify the probability of the input feature vector x supplemented to the NN belonging to one of two classes {−1, +1}. The output of the NN will be a probability (i.e. a value ranging between 0 and 1) that indicates the probability of the input belonging to class +1. The probability of belonging to the other class −1 will be 1 minus the output probability. The NN with L layers will be composed of the input, output and hidden layers as follow:

The input layer is composed of d + 1 inputs where d is the number of features in each input vector x and a bias node.

The output layer is composed of a single node which outputs the probability of the input vector x belonging to class +1.